+++ common/kafka/src/test/java/com/f4/fqs/commons/domain/DomainApplicationTests.java
