+++ common/kafka_common/build.gradle
@@ -1,0 +1,25 @@
+plugins {
+	id 'java'
+}
+
+dependencies {
+
+	implementation project(':common:domain')
+
+	implementation 'org.springframework.boot:spring-boot-starter'
+	implementation 'com.fasterxml.jackson.core:jackson-databind:_'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+//	implementation 'org.springframework.boot:spring-boot-starter-webflux'
+	implementation 'org.springframework.kafka:spring-kafka'
+//	implementation 'io.projectreactor.kafka:reactor-kafka:1.3.22'
+	testImplementation 'org.springframework.kafka:spring-kafka-test'
+//	testImplementation 'io.projectreactor:reactor-test'
+}
+
+tasks.named('test') {
+	useJUnitPlatform()
+}
+
+tasks.bootJar { enabled = false }
