+++ common/kafka_common/src/main/java/com/f4/fqs/commons/kafka_common/exception/BusinessException.java
@@ -1,0 +1,15 @@
+package com.f4.fqs.commons.kafka_common.exception;
+
+import com.f4.fqs.commons.domain.exception.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
