+++ common/kafka_common/src/main/java/com/f4/fqs/commons/kafka_common/util/CommonConstraints.java
@@ -1,0 +1,10 @@
+package com.f4.fqs.commons.kafka_common.util;
+
+public final class CommonConstraints {
+
+    public static final String KAFKA_TOPIC_EVENT_PREFIX = "EVENT_";
+    public static final String KAFKA_TOPIC_QUEUE_PREFIX = "QUEUE_";
+    public static final String QUEUE_NAME = "queue";
+    public static final String EVENT = "event";
+
+}
