+++ common/kafka_common/src/test/java/com/f4/fqs/commons/domain/DomainApplicationTests.java
@@ -1,0 +1,13 @@
+package com.f4.fqs.commons.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DomainApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
