+++ common/reactive_kafka/src/main/java/com/f4/fqs/commons/reactive_kafka/kafka/config/ComponentScanConfig.java
@@ -1,0 +1,9 @@
+package com.f4.fqs.commons.reactive_kafka.kafka.config;
+
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan(basePackages = {
+        "com.f4.fqs.commons.kafka_common"
+})
+public class ComponentScanConfig {
+}
