+++ common/reactive_kafka/src/main/java/com/f4/fqs/commons/reactive_kafka/kafka/producer/ProducerService.java
@@ -1,0 +1,10 @@
+package com.f4.fqs.commons.reactive_kafka.kafka.producer;
+
+import com.f4.fqs.commons.kafka_common.message.QueueCommand;
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+public interface ProducerService {
+
+    void send(String topic, QueueCommand message) throws JsonProcessingException;
+
+}
