--- common/store/src/main/java/com/f4/fqs/commons/store/exception/BusinessException.java
@@ -1,15 +1,0 @@
-package com.f4.fqs.commons.store.exception;
-
-import com.f4.fqs.commons.domain.exception.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class BusinessException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
