--- common/store/src/main/java/com/f4/fqs/commons/store/kafka/producer/EventProducerService.java
@@ -1,29 +1,0 @@
-package com.f4.fqs.commons.store.kafka.producer;
-
-import com.f4.fqs.commons.store.message.QueueCommand;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.kafka.core.reactive.ReactiveKafkaProducerTemplate;
-import org.springframework.stereotype.Service;
-
-import java.util.Objects;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public final class EventProducerService implements ProducerService {
-
-    private final ReactiveKafkaProducerTemplate<String, QueueCommand> producer;
-
-    @Override
-    public void send(String topic, QueueCommand message) {
-
-        producer.send(topic, message).doOnError(e -> {
-            if (!Objects.isNull(e)) {
-                log.error("이벤트 생성 에러 발생 :: {}", e.getMessage());
-            }
-        })
-        .subscribe();
-    }
-
-}
