+++ service/event_store/src/main/java/com/f4/fqs/eventStore/kafka/config/KafkaConfig.java
@@ -1,7 +1,7 @@
 package com.f4.fqs.eventStore.kafka.config;
 
 
-import com.f4.fqs.commons.store.message.QueueCommand;
+import com.f4.fqs.commons.kafka_common.message.QueueCommand;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,7 +11,6 @@ import org.springframework.kafka.annotation.EnableKafka;
 import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
 import org.springframework.kafka.core.ConsumerFactory;
 import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-import org.springframework.kafka.core.reactive.ReactiveKafkaConsumerTemplate;
 import org.springframework.kafka.listener.ContainerProperties;
 import org.springframework.kafka.support.serializer.JsonDeserializer;
