+++ service/queue/build.gradle
@@ -9,7 +9,8 @@ ext {
 
 dependencies {
 	implementation project(':common:domain')
-	implementation project(':common:store')
+	implementation project(':common:reactive_kafka')
+	implementation project(':common:kafka_common')
 
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
