+++ service/queue_manage/build.gradle
@@ -9,6 +9,8 @@ ext {
 
 dependencies {
     implementation project(':common:domain')
+    implementation project(':common:kafka_common')
+    implementation project(':common:kafka')
     implementation project(':service:gateway:gateway_domain')
 
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
@@ -17,6 +19,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
+    implementation 'org.springframework.kafka:spring-kafka'
 //    implementation 'com.github.docker-java:docker-java:3.0.6'
     runtimeOnly 'com.mysql:mysql-connector-j'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
