+++ service/queue_manage/src/main/java/com/f4/fqs/queue_manage/application/service/KafkaService.java
@@ -1,0 +1,23 @@
+package com.f4.fqs.queue_manage.application.service;
+
+import com.f4.fqs.commons.kafka.producer.EventSourcingExecutor;
+import com.f4.fqs.commons.kafka_common.message.QueueCommand;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class KafkaService {
+
+    private final EventSourcingExecutor executor;
+
+    public void test() {
+        executor.createEvent(QueueCommand.addQueueCommand("test-service", UUID.randomUUID(), LocalDateTime.now()));
+    }
+
+}
