+++ service/queue_manage/src/main/resources/application.yml
@@ -20,6 +20,11 @@ spring:
         default_batch_fetch_size: 100
     open-in-view: false
     show-sql: true
+  kafka:
+    bootstrap-servers: localhost:9092
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
 
   data:
     redis:
