+++ service/user/build.gradle
@@ -16,6 +16,7 @@ dependencies {
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
     runtimeOnly 'com.mysql:mysql-connector-j'
     implementation project(":common:domain")
+    implementation project(":common:kafka")
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
