+++ service/user/src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
 
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:5432/fqs
+    url: jdbc:mysql://localhost:3306/fqs
     username: root
     password: password
