+++ settings.gradle
@@ -15,4 +15,6 @@ include 'service:payment'
 include 'service:event_store'
 
 include 'common:domain'
-include 'common:store'
+include 'common:kafka'
+include 'common:reactive_kafka'
+include 'common:kafka_common'
